+++ lib/src/analyzers/unused_l10n_analyzer/unused_l10n_analyzer.dart
@@ -166,8 +166,8 @@ class UnusedL10nAnalyzer {
     CompilationUnitElement unit,
   ) {
     final unusedAccessors = classElement.accessors
-        .where((field) => !usages.contains(field.name) && !field.isSynthetic)
-        .toList();
+        .where((field) => !usages.contains(field.name))
+        .map((field) => field.isSynthetic ? field.nonSynthetic : field);
 
     return unusedAccessors
         .map((accessor) => _createL10nIssues(accessor as ElementImpl, unit))
@@ -180,10 +180,8 @@ class UnusedL10nAnalyzer {
     CompilationUnitElement unit,
   ) {
     final unusedMethods = classElement.methods
-        .where(
-          (method) => !usages.contains(method.name) && !method.isSynthetic,
-        )
-        .toList();
+        .where((method) => !usages.contains(method.name))
+        .map((method) => method.isSynthetic ? method.nonSynthetic : method);
 
     return unusedMethods
         .map((method) => _createL10nIssues(method as ElementImpl, unit))