+++ test/resources/unused_l10n_analyzer/unused_l10n_analyzer_example.dart
@@ -2,6 +2,8 @@ import 'test_i18n.dart';
 
 void main() {
   final _ = TestI18n.method('value');
+
+  S.getter;
 }
 
 class SomeClass {